backend_requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography>=41.0.0,<46.0.0
cymem>=2.0.8,<3.0.0
ecdsa>=0.18.0,<0.20.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy>=1.24.0,<3.0.0
packaging>=20.0,<25.0
passlib>=1.7.0,<2.0.0
preshed==3.0.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich>=10.14.0,<14.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.3
thinc==8.3.6
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3

# Frontend charts - retained for potential server-side generation
plotly==5.24.1
rank-bm25==0.2.2
readability-lxml==0.8.1
lxml[html_clean]==5.3.0
openai==1.51.2
httpx==0.27.2

# Database dependencies - for backend only
psycopg2-binary==2.9.9

